build: split `--verbose` to specific flags
Usage (excluding initialization) of `build.verbose` appears in two
files:

1. lib/std/Build.zig
2. lib/std/Build/Step.zig

As it turns out only Step.zig is relevant in the new context of
`--verbose-spawn`. The verbose references in Build.zig are through
`makeUninstall` and `truncateFile`. The first of which is a bunch of
deletion operations, and the second of which only got called from
InstallDir.zig.

The `truncateFile` operation in itself is general enough to not depend
on a specific flag and the log call is therefore hoisted to the
InstallDir.zig callsite. The logging in InstallDir is then enabled by
a new parameter `--verbose-install`. I also took the liberty of adding
logging to the result of updateFile operations as well under this flag.
The updateFile logic is also used in InstallFile and InstallArtifact, so
the helper `handleVerboseInstallUpdateFile` was created and added to all
call site results.

Regarding audit of child process invocations in Build.zig and default
Steps:

`execAllowFail` is the child process spawning part of the Build API that
did not log invocations, so a call to `handleVerboseSpawn2` was added.

The other APIs are `evalZigProcess` and `evalChildProcess`, but they
already use the `handleVerboseSpawn` helper.

Closes #14970