std.Uri: Assume query string is already escaped
std.Uri.format currently calls writeEscapedQuery when writing out the
query string. Unfortunately, this means that it is not possible to
specify a query that contains an ampersand or an equals sign, since
these are left unescaped by the writeEscapedQuery.

For example, assume you are attempting to pass the value `M&M` as a
query variable named `candy`, and the value `=` as the variable `cmp`.
Note that the escaped value for `=` is `%3D`, `&` is `%26', and '%' is
'%25'.

Consider the possible values for Uri.query attempting to encode the
above example:
1. Unescaped: `candy=M&M&cmp==`
2. Escaped: `candy=M%26M&cmp=%3D`

After passing through writeEscapedQuery, (1) encodes to itself, because
`=` and `&` are valid characters in a query string. However, the
semantic meaning of that string unclear, but at minimum contains the set
`{ candy: "M", M: "" }`, and the behavior around `cmp` is dependent on
any rules that may or may not exist for interpreting malformed query
values.

On the other hand, (2) would end up double-escaped because '%' gets
escaped to '%25', resulting in `candy=M%2526M^cmp=%253D`, which is the
set `{ candy: "M%26", cmp: "%3D"}`, which is also wrong.

Ultimately, if there's any instances of `=` or `&` in the query
variable, not only is writeEscapedQuery incorrect, but there's no way to
write a correct function that operates on a combined query string,
rather than a set of key/value pairs.

Instead, Uri should assume the query string is already escaped. Callers
can use Uri.escapeString and Uri.writeEscapedString to encode their
query values individually, and then combine.

This approach is similar to the Go standard library, which uses a
URL.RawQuery field to store the encoded query as a string.