std.Progress: Do not use escape codes on Windows even if support for them is detected
The particular escape code sequences used by the progress bar do not seem to get parsed if the code page is set to 437 (the default code page for a console session), and even if the escape sequences are parsed, the Windows implementation does not render them correctly (it only partially clears the line so the output just ends up mangled).

Until this is more understood, we just force the Windows Console APIs to be used on Windows no matter what.

See #16526