wasm: do not default to pic and re-enable tests
For WebAssembly we want DynamicNoPIC for relocation mode. PIC is experimental,
but will also force a shared library for WebAssembly which is not what the user
expects as that will require runtime relocations to be performed. This also means
that we generate an invalid object file unless Zig (or the user) appends `-shared`
to wasm-ld.

This also re-enables the linker tests that regressed with llvm17