std.BoundedArray: make capacity a comptime field
I think the more compatible `std.ArrayList` and `std.BoundedArray` are,
the better.
This makes the two slightly more compatible because `std.ArrayList` also
has its capacity as a field.
In the case of `std.BoundedArray` it's going to be a comptime field and so doesn't make
the data structure any bigger.

The main motivation for this however was that there's not actually any
way to get the capacity of a `std.BoundedArray` at comptime which makes it
difficult for example to easily make use of the passed capacity when defining
global variables for example without making the capacity I pass another
constant beforehand, or doing some weird comptime/`@TypeOf` stuff.