Extend std.os.waitpid and wait4 to also return ECHILD
Per the man pages for waitpid calling waitpid with pid = 0 will wait for
any child process belonging to the process group, where pid = -1, will
wait for any child process. This isn't always a race condition, and
swallowing this error code with a panic & crash precludes a number of
designs. Waiting for a child process that the waiting process doesn't
already know about is perfectly valid, and in some cases required.

This modifies the API and allows `std.os.waitpid`, and `std.os.wait4` to
also return a singular error state for ECHILD.