build: don't skip compiler build with -Dno-bin
This option exists for fast iteration during compiler development, since
avoiding codegen (but still performing semantic analysis) causes compile
errors to appear around twice as fast. This option was broken when the
`emit_bin` property on `std.Build.Step.Compile` was removed.

Now, when this option is passed, we add a direct install dependency on
the compile step itself, so that it is always run.